% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjRandIndex.R
\name{AdjRandIndex}
\alias{AdjRandIndex}
\title{Calculate Adjusted Rand Index}
\usage{
AdjRandIndex(x, y)
}
\arguments{
\item{x}{A vector of class labels}

\item{y}{A vector of class labels to be compared with \code{x}. The length of \code{y} should be the same as \code{x}.}
}
\value{
This function returns Adjusted Rand Index between two label vectors \code{x} and \code{y}.
}
\description{
Calculate Adjusted Rand Index measuring the similarity between two clustering labels.
}
\examples{
# true label
x <- rep(1:3, each = 10)

# predicted label
y <- sample(x)

# compare x and y
AdjRandIndex(x, y)
NormMI(x, y)
Purity(x, y)
}
\references{
\insertRef{hubert1985comparing}{MKKC}
}
\seealso{
\code{\link{NormMI}}, \code{\link{Purity}}
}
