% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkkcEst.R
\name{mkkcEst}
\alias{mkkcEst}
\title{An internal function called \code{mkkcEst}.}
\usage{
mkkcEst(K, centers, iter.max = 10, A = NULL, bc = NULL,
  epsilon = 1e-04, theta = rep(1/dim(Km)[3], dim(Km)[3]))
}
\arguments{
\item{K}{\eqn{N x N x P} array containing \eqn{P} kernel matrices with size \eqn{N x N}.}

\item{centers}{The number of clusters, say \eqn{k}.}

\item{iter.max}{The maximum number of iterations allowed. The default is 10.}

\item{A}{Linear constraint matrix.}

\item{bc}{Lower and upper constraint bounds.}

\item{epsilon}{Convergence threshold. The default is \eqn{10^{-4}}.}

\item{theta}{intial values for kernel coefficients. The default is 1/P for all views.}
}
\value{
\code{mkkcEst} returns the following components:
\describe{
  \item{cluster}{A vector of integers (from \code{1:k}) indicating the cluster to which each point is allocated.}
  \item{totss}{The total sum of squares.}
  \item{withinss}{Matrix of within-cluster sum of squares by cluster, one row per view.}
  \item{withinsscluster}{Vector of within-cluster sum of squares, one component per cluster.}
  \item{withinssview}{Vector of within-cluster sum of squares, one component per view.}
  \item{tot.withinss}{Total within-cluster sum of squares, i.e. \code{sum(withinsscluster)}.}
  \item{betweenssview}{Vector of  between-cluster sum of squares, one component per view.}
  \item{tot.betweenss}{The between-cluster sum of squares, i.e. \code{totss-tot.withinss}.}
  \item{clustercount}{The number of clusters, say \code{k}.}
  \item{coefficients}{The kernel coefficients}
  \item{size}{The number of points, one component per cluster.}
  \item{iter}{The number of iterations.}
}
}
\description{
It performs multiple kernel K-means clustering on a multi-view data.
}
\keyword{internal}
