% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormMI.R
\name{NormMI}
\alias{NormMI}
\alias{MI}
\alias{Entropy}
\title{Calculate Normalized Mutual Information}
\usage{
NormMI(x, y)

MI(x, y)

Entropy(x)
}
\arguments{
\item{x}{A vector of class labels.}

\item{y}{A vector of class labels to be compared with \code{x}. The length of \code{y} should be the same as \code{x}.}
}
\value{
\code{NormMI} returns Normalized Mutual Information between two label vectors \code{x} and \code{y}.

\code{MI} returns Mutual Information between two label vectors.

\code{Entropy} returns Entropy of the label vector \code{x}.
}
\description{
Calculate Normalized Mutual Information measuring the mutual dependence between two clustering labels.
}
\examples{
# true label
x <- rep(1:3, each = 10)

# predicted label
y <- sample(x)

# compare x and y
AdjRandIndex(x, y)
NormMI(x, y)
Purity(x, y)
}
\references{
\insertRef{strehl2002cluster}{MKKC}
}
\seealso{
\code{\link{AdjRandIndex}}, \code{\link{Purity}}
}
