% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Purity.R
\name{Purity}
\alias{Purity}
\title{Calculate Purity}
\usage{
Purity(x, y)
}
\arguments{
\item{x}{A vector of the predicted class labels}

\item{y}{A vector of the true class labels. The length of \code{y} should be the same as \code{x}.}
}
\value{
This function returns Purity between two label vectors \code{x} and \code{y}.
}
\description{
Calculate Purity measuring the mutual dependence between two clustering labels.
}
\examples{
# true label
x <- rep(1:3, each = 10)

# predicted label
y <- sample(x)

# compare x and y
AdjRandIndex(x, y)
NormMI(x, y)
Purity(x, y)
}
\references{
\insertRef{Manning2008information}{MKKC}
}
\seealso{
\code{\link{AdjRandIndex}}, \code{\link{NormMI}}
}
