% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardizeKernel.R
\name{StandardizeKernel}
\alias{StandardizeKernel}
\title{Standardize a Kernel Matrix}
\usage{
StandardizeKernel(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{A kernel matrix to be standardized.}

\item{center}{Logical. If \code{TRUE}, the kernel matrix \code{x} will be centered. The default is \code{TRUE}}

\item{scale}{Logical. If \code{TRUE}, the kernel matrix \code{x} will be scaled. The default is \code{TRUE}.}
}
\value{
This function returns the standardized kernel matrix. If both arguments \code{center} and \code{scale} are \code{TRUE}, the kernel matrix will be centered first and then scaled.
}
\description{
\code{StandardizeKernel} is used to standardize kernel matrices obtained from multiple views so that the views are comparable to each other. A kernel matrix \code{x} is centered and then scaled by dividing by its trace.
}
\examples{
x <- diag(3); x
StandardizeKernel(x, center = TRUE, scale = FALSE)   # centered
StandardizeKernel(x, center = FALSE, scale = TRUE)   # scaled
StandardizeKernel(x, center = TRUE, scale = TRUE)    # centered and scaled
}
\references{
\insertRef{scholkopf1998nonlinear}{MKKC}

\insertRef{bang2018mkkc}{MKKC}
}
